+++ api/src/certification/enrolment/application/enrolment-route.js
@@ -1,6 +1,7 @@
 import JoiDate from '@joi/date';
 import BaseJoi from 'joi';
 const Joi = BaseJoi.extend(JoiDate);
+import { sessionController } from '../../../../lib/application/sessions/session-controller.js';
 import { identifiersType } from '../../../shared/domain/types/identifiers-type.js';
 import { authorization } from '../../shared/application/pre-handlers/authorization.js';
 import { enrolmentController } from './enrolment-controller.js';
@@ -50,6 +51,33 @@ const register = async function (server) {
         notes: [
           '- **Cette route est restreinte aux utilisateurs appartenant à un centre de certification ayant créé la session**\n' +
             "- Cette route permet de télécharger le template d'import des candidats d'une certification au format ods",
+        ],
+      },
+    },
+    {
+      method: 'POST',
+      path: '/api/sessions/{id}/certification-candidates/import',
+      config: {
+        validate: {
+          params: Joi.object({
+            id: identifiersType.sessionId,
+          }),
+        },
+        payload: {
+          parse: 'gunzip',
+          maxBytes: 1048576 * 10, // 10MB
+        },
+        pre: [
+          {
+            method: authorization.verifySessionAuthorization,
+            assign: 'authorizationCheck',
+          },
+        ],
+        handler: sessionController.importCertificationCandidatesFromCandidatesImportSheet,
+        tags: ['api', 'sessions'],
+        notes: [
+          '- **Cette route est restreinte aux utilisateurs authentifiés et appartenant à un centre de certification ayant créé la session**\n' +
+            '- Elle permet de récupérer la liste des candidats à inscrire contenue dans le PV de session format ODS envoyé',
         ],
       },
     },
