+++ api/tests/unit/application/session/index_test.js
@@ -1,10 +1,3 @@
-import fs from 'node:fs';
-import { stat, unlink, writeFile } from 'node:fs/promises';
-import * as url from 'node:url';
-
-import FormData from 'form-data';
-import streamToPromise from 'stream-to-promise';
-
 import { NotFoundError } from '../../../../lib/application/http-errors.js';
 import * as moduleUnderTest from '../../../../lib/application/sessions/index.js';
 import { sessionController } from '../../../../lib/application/sessions/session-controller.js';
@@ -13,82 +6,7 @@ import { authorization } from '../../../../src/certification/shared/application/
 import { securityPreHandlers } from '../../../../src/shared/application/security-pre-handlers.js';
 import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
 
-const __dirname = url.fileURLToPath(new URL('.', import.meta.url));
-
 describe('Unit | Application | Sessions | Routes', function () {
-  describe('POST /api/sessions/{id}/certification-candidates/import', function () {
-    const testFilePath = `${__dirname}/testFile_temp.ods`;
-    const method = 'POST';
-
-    let headers;
-    let payload;
-
-    beforeEach(async function () {
-      await writeFile(testFilePath, Buffer.alloc(0));
-      const form = new FormData();
-      const knownLength = await stat(testFilePath).size;
-      form.append('file', fs.createReadStream(testFilePath), { knownLength });
-
-      headers = form.getHeaders();
-      payload = await streamToPromise(form);
-    });
-
-    afterEach(async function () {
-      await unlink(testFilePath);
-    });
-
-    it('should exist', async function () {
-      // given
-      sinon.stub(authorization, 'verifySessionAuthorization').returns(null);
-      sinon.stub(sessionController, 'importCertificationCandidatesFromCandidatesImportSheet').returns('ok');
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      const sessionId = 3;
-      const url = `/api/sessions/${sessionId}/certification-candidates/import`;
-
-      // when
-      const response = await httpTestServer.request(method, url, payload, null, headers);
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-
-    context('when session ID params is not a number', function () {
-      it('should return 400', async function () {
-        // given
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-
-        const sessionId = 'salut';
-        const url = `/api/sessions/${sessionId}/certification-candidates/import`;
-
-        // when
-        const response = await httpTestServer.request(method, url, payload, null, headers);
-
-        // then
-        expect(response.statusCode).to.equal(400);
-      });
-    });
-
-    context('when session ID params is out of range for database integer (> 2147483647)', function () {
-      it('should return 400', async function () {
-        // given
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-
-        const sessionId = 9999999999;
-        const url = `/api/sessions/${sessionId}/certification-candidates/import`;
-
-        // when
-        const response = await httpTestServer.request(method, url, payload, null, headers);
-
-        // then
-        expect(response.statusCode).to.equal(400);
-      });
-    });
-  });
-
   describe('POST /api/sessions/{id}/candidate-participation', function () {
     it('should exist', async function () {
       // given
